ParseLog/requirements.txt
msgpack-python==0.4.7
ujson==1.35
